// infrastructure/kubernetes/generatedCode/crds/argoproj/v1alpha1/workflow.ts
// *** WARNING: this file was generated by crd2pulumi. ***
// *** Do not edit by hand unless you're certain you know what you are doing! ***

import * as pulumi from "@pulumi/pulumi";
import { input as inputs, output as outputs } from "../../types";
import * as utilities from "../../utilities";

import { ObjectMeta } from "../../meta/v1";

export class Workflow extends pulumi.CustomResource {
	/**
	 * Get an existing Workflow resource's state with the given name, ID, and optional extra
	 * properties used to qualify the lookup.
	 *
	 * @param name The _unique_ name of the resulting resource.
	 * @param id The _unique_ provider ID of the resource to lookup.
	 * @param opts Optional settings to control the behavior of the CustomResource.
	 */
	public static get(
		name: string,
		id: pulumi.Input<pulumi.ID>,
		opts?: pulumi.CustomResourceOptions,
	): Workflow {
		return new Workflow(name, undefined as any, { ...opts, id: id });
	}

	/** @internal */
	public static readonly __pulumiType =
		"kubernetes:argoproj.io/v1alpha1:Workflow";

	/**
	 * Returns true if the given object is an instance of Workflow.  This is designed to work even
	 * when multiple copies of the Pulumi SDK have been loaded into the same process.
	 */
	public static isInstance(obj: any): obj is Workflow {
		if (obj === undefined || obj === null) {
			return false;
		}
		return obj["__pulumiType"] === Workflow.__pulumiType;
	}

	public readonly apiVersion!: pulumi.Output<
		"argoproj.io/v1alpha1" | undefined
	>;
	public readonly kind!: pulumi.Output<"Workflow" | undefined>;
	public readonly metadata!: pulumi.Output<ObjectMeta>;
	public readonly spec!: pulumi.Output<outputs.argoproj.v1alpha1.WorkflowSpec>;
	public readonly status!: pulumi.Output<
		outputs.argoproj.v1alpha1.WorkflowStatus | undefined
	>;

	/**
	 * Create a Workflow resource with the given unique name, arguments, and options.
	 *
	 * @param name The _unique_ name of the resource.
	 * @param args The arguments to use to populate this resource's properties.
	 * @param opts A bag of options that control this resource's behavior.
	 */
	constructor(
		name: string,
		args?: WorkflowArgs,
		opts?: pulumi.CustomResourceOptions,
	) {
		let resourceInputs: pulumi.Inputs = {};
		opts = opts || {};
		if (!opts.id) {
			resourceInputs["apiVersion"] = "argoproj.io/v1alpha1";
			resourceInputs["kind"] = "Workflow";
			resourceInputs["metadata"] = args ? args.metadata : undefined;
			resourceInputs["spec"] = args ? args.spec : undefined;
			resourceInputs["status"] = args ? args.status : undefined;
		} else {
			resourceInputs["apiVersion"] = undefined /*out*/;
			resourceInputs["kind"] = undefined /*out*/;
			resourceInputs["metadata"] = undefined /*out*/;
			resourceInputs["spec"] = undefined /*out*/;
			resourceInputs["status"] = undefined /*out*/;
		}
		opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
		super(Workflow.__pulumiType, name, resourceInputs, opts);
	}
}

/**
 * The set of arguments for constructing a Workflow resource.
 */
export interface WorkflowArgs {
	apiVersion?: pulumi.Input<"argoproj.io/v1alpha1">;
	kind?: pulumi.Input<"Workflow">;
	metadata?: pulumi.Input<ObjectMeta>;
	spec?: pulumi.Input<inputs.argoproj.v1alpha1.WorkflowSpecArgs>;
	status?: pulumi.Input<inputs.argoproj.v1alpha1.WorkflowStatusArgs>;
}